+++ tests/replicated_graph_test.py
@@ -140,3 +140,73 @@ def test_too_many_ipus(trace_model):
     ):
         poptorch_model(torch.tensor(input, requires_grad=True),
                        torch.tensor(labels))
+
+
+class ModelWithLoss(torch.nn.Module):
+    def __init__(self, W_init):
+        super().__init__()
+        self.W = torch.nn.Parameter(W_init)
+
+    def forward(self, X):
+        Z = X @ self.W
+        return Z, poptorch.identity_loss(Z**2, reduction="mean")
+
+
+@pytest.mark.ipuHardwareRequired
+def test_per_replica_variables():
+    # Split the weight tensor into 4, and the input data tensor into 2.
+    tensor_shards = 4
+    data_shards = 2
+
+    # Set up the problem
+    random = np.random.RandomState(seed=100)
+    prob_X = random.normal(size=(24, 40)).astype(np.float32)
+    prob_W_init = random.normal(size=(40, 56)).astype(
+        np.float32) * (5 * 8)**-0.5
+    prob_steps = 4
+
+    # Run on the CPU
+    X = torch.tensor(prob_X)
+    W = torch.nn.Parameter(torch.tensor(prob_W_init))
+    optim = torch.optim.SGD([W], lr=0.01)
+
+    cpu_losses = []
+    for _ in range(prob_steps):
+        optim.zero_grad()
+        v = (X @ W)**2
+        loss = torch.mean(v)
+        loss.backward()
+        optim.step()
+        cpu_losses.append(loss.detach())
+    cpu_losses = np.array(cpu_losses)
+    cpu_W_final = W.detach().numpy()
+
+    # Run on 8 IPUs
+    W_init = torch.tensor(
+        prob_W_init.reshape(prob_W_init.shape[0], tensor_shards,
+                            prob_W_init.shape[1] // tensor_shards).transpose(
+                                1, 0, 2)).contiguous()
+    m = ModelWithLoss(W_init)
+    optim = torch.optim.SGD(m.parameters(), lr=0.01)
+
+    pt_opts = poptorch.Options()
+    pt_opts.replicationFactor(data_shards * tensor_shards, data_shards)
+    pt_opts.outputMode(poptorch.OutputMode.All)
+    pt_m = poptorch.trainingModel(m, optimizer=optim, options=pt_opts)
+    pt_m.W.perReplica(poptorch.enums.CommGroupType.Orthogonal, data_shards,
+                      poptorch.enums.VariableRetrievalMode.OnePerGroup)
+    pt_losses = []
+    if data_shards > 1:
+        X = X.reshape(data_shards, X.shape[0] // data_shards, *X.shape[1:])
+    for _ in range(prob_steps):
+        _, loss = pt_m(X)
+        # We divide by the number of replicas because the mean is being
+        # taken only over a part of the tensor on each replica, so we need to
+        # divide by the number of replicas to get the correct mean.
+        pt_losses.append(
+            torch.sum(loss.detach()) / (data_shards * tensor_shards))
+    pt_losses = np.array(pt_losses)
+    pt_W_final = m.W.detach().numpy().transpose(1, 0, 2) \
+                  .reshape(prob_W_init.shape)
+    np.testing.assert_allclose(cpu_losses, pt_losses, atol=1e-6)
+    np.testing.assert_allclose(cpu_W_final, pt_W_final, atol=1e-6)